Thomas Lundquist writes:
If I understand you correctly, you want the ncurses development headers
on the target.

a patch for this (named target_headers.patch and includes similar
options for a few other libs in buildroot) can be found at:

http://www.zelow.no/floppyfw/download/Development/Patches/buildroot/

(a few packages there aswell)

it will add an option to put headers on target for ncurses, zlib and
openssl.



Thomas.