pkgconfig: Return appropriate error code from pkgconfig script
Patch by Will Newton.

Currently the wrapper script for pkgconfig clobbers the return status
of pkgconfig. This is a problem for example when someone tries to run
"pkgconfig --exists". This patch should allow the correct return code
to be passed through. Note it uses the "pipefail" option that was
introduced with bash 3.