The liboil Makefile will use "-mcpu=vfp" when building for ARM. The VFP is only available for a few ARM CPUs at the moment, so this breaks the liboil build.
A patch is available upstream which only enables "-mfpu=vfp"
if "--enable-vfp" is given to "configure".
Autotools needs to be run for liboil for this to take effect.

A new configuration BR2_VFP_FLOAT is added to allow enabling vfp.
If this is "yes", then "-mfpu=vfp" is added to CFLAGS.