avahi-autoipd: fix issue with repeated builds
The avahi-autoipd target install step adds /var/lib/avahi-autoipd as a
symlink to /tmp/avahi-autoipd, which is fine, except if your /var/lib is
already a symlink to /tmp (E.G. like the default target_skeleton is).

If that's the case, then you end up adding a /tmp/avahi-autoipd symlink
pointing to itself, causing subsequent builds to fail
(Too many levels of symbolic links in the ln -sf call). Fix it by simply
first removing the symlink if it's there.