package/qt: add font selection
Commit f98547622 (make sure to install all Qt-fonts and true-type-fonts)
changed Qt to install all .qpf fonts to the target instead of only a
hardcoded subset. Some of those fonts are unfortunately quite big (1-2MB)
and are not always needed (japanese, unicode).

Instead add a font selection in Kconfig similar to how it is done for
pixel depths, and default to the subset we were previously using.

Signed-off-by: Peter Korsgaard <jacmet@sunsite.dk>