dependencies: bail out if makeinfo isn't found
makeinfo is mandatory at least to build a toolchain. Currently,
dependencies.sh doesn't fail if makeinfo is not present and only
displays a message that can easily be lost. The user will then
encounter the issue later, when it is more difficult to understand
what's happening.

So, this patch simply does for makeinfo what the script does for the
other dependencies: bail out if they aren't available.

Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>