libxml-parser-perl: new package
This Perl module is needed to build intltool, which in turn is needed
to build several of our packages. We assume that Perl is installed on
the host.

Only the host variant of this package has been tested and is used. The
code for the target variant (what to do at the configure, build and
install steps) isn't implemented, but the GENTARGETS call for the
target is mandatory, due to limitations of the Makefile.package.in
infrastructure (if one wants the host package to be implemented, then
the target package *must* exist). This limitation will be removed
after 2010.02.

Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>