libsoup: remove --without-html-dir
--without-html-dir doesn't work. It leads libsoup ./configure script
to think that HTML_DIR is "no", which leads to the installation of the
documentation in $(STAGING_DIR)no (yes, with the "no" suffix at the
end). The issue is that --with-html-dir is not an enable/disable type
of option, it's an option that only allows to pass a PATH for
documentation installation.

As we don't want the documentation to be installed in this odd
$(STAGING_DIR)no directory, we simply get rid of this option.

See 7e62e8ec723f91e07bb4ac9ae3e55b65e3123082 for another instance of
the same problem.

Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>