remove -mm linux patch options
The -mm tree is really no longer available since the -next tree
was created for testing.  It's not even listed on the kernel.org
site any longer.  Plus, the last one that was published as
available was pretty old, 2.6.28-rc2-mm1.

Remove all the -mm patch stuff from Buildroot to clean up the
Linux kernel selection.

Signed-off-by: H Hartley Sweeten <hsweeten@visionengravers.com>
Signed-off-by: Peter Korsgaard <jacmet@sunsite.dk>