usb_modeswitch: use GENTARGETS instead of AUTOTARGETS
usb_modeswitch build system is just a simple Makefile, so the
GENTARGETS infrastructure is more appropriate than the AUTOTARGETS
infrastructure. The build system is slightly fixed through a small
patch, that replaces the previous "nostrip" patch.

The latest version of usb_modeswitch is 1.1.1, but it now relies on
udev and requires TCL to operate. Someone actually using
usb_modeswitch would probably be at a better place to do this version
bump.

Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>