New graph-depends script
Add a Python script that draws dependency graphs, either for the full
set of selected packages, or for a particular given package. Of
course, it is limited to packages that use either the generic or
autotools infrastructures. The script generates a file that Graphviz
can parse and generate a PDF (or other formats) from it.

Full dependency graph of all selected packages :

 ./scripts/graph-depends > test.dot
 dot -Tpdf test.dot -o test.pdf

Dependency graph of libgtk2 :

 ./scripts/graph-depends libgtk2 > test.dot
 dot -Tpdf test.dot -o test.pdf

Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>