Make uClibc gen_wc8bit shows an error when no locale support available
When no UTF-8 locale is available on the host system, uClibc can't
generate some stuff it needs to compile a C library with locale
support. Unfortunately, as gen_wc8bit message is shown on stdout and
the stdout of gen_wc8bit is redirected to a file, the user don't see
anything, as reported at
http://lists.busybox.net/pipermail/buildroot/2010-May/034177.html.

Those two patches fix the problem for uClibc 0.9.31 and 0.9.30.3. It
has been submitted upstream:
 http://lists.uclibc.org/pipermail/uclibc/2010-August/044256.html

Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>