board: remove i386/i686_defconfig
Those are not associated with any specific hardware system (PC or
another i386 system). Moreover, the fact that those configurations
require the build of a JFFS2 filesystem, very uncommon on PC systems,
seems to indicate that those configurations are not really being used
today.

It would make more sense to have a qemu_i388_defconfig (building a
kernel with just the device drivers for Qemu) and possibly a
pc_i386_defconfig (building a kernel with many device drivers, and a
bootloader such as Grub or Grub 2).

We also remove the corresponding kernel configuration files.

Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>
Signed-off-by: Peter Korsgaard <jacmet@sunsite.dk>